examples/python/pyproject.toml
[project]
name = "python"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "anyio>=4.12.1",
    "googleapis-common-protos>=1.72.0",
    "grpcio>=1.78.0",
    "grpcio-reflection>=1.78.0",
    "mcp>=1.26.0",
    "protobuf>=6.33.5",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
